Rename `commandHandler` package to `command`
