+++ src/main/java/command/CommandHandler.java
@@ -1,4 +1,4 @@
-package commandhandler;
+package command;
 
 import data.TaskList;
 import java.util.List;
