+++ src/main/java/command/CommandMarkHandler.java
@@ -1,4 +1,4 @@
-package commandhandler;
+package command;
 
 import data.TaskList;
 import data.tasks.Task;
